:recycle: api: GET /api/sessions/{id}/certification-candidates to src
