+++ api/lib/application/sessions/index.js
@@ -178,29 +178,6 @@ const register = async function (server) {
       },
     },
     {
-      method: 'GET',
-      path: '/api/sessions/{id}/certification-candidates',
-      config: {
-        validate: {
-          params: Joi.object({
-            id: identifiersType.sessionId,
-          }),
-        },
-        pre: [
-          {
-            method: authorization.verifySessionAuthorization,
-            assign: 'authorizationCheck',
-          },
-        ],
-        handler: sessionController.getCertificationCandidates,
-        tags: ['api', 'sessions', 'certification-candidates'],
-        notes: [
-          'Cette route est restreinte aux utilisateurs authentifiés',
-          'Elle retourne les candidats de certification inscrits à la session.',
-        ],
-      },
-    },
-    {
       method: 'DELETE',
       path: '/api/sessions/{id}/certification-candidates/{certificationCandidateId}',
       config: {
