+++ api/tests/certification/session/unit/application/certification-candidate-route_test.js
@@ -121,6 +121,14 @@ describe('Unit | Application | Sessions | Routes', function () {
         condition: 'session ID params is out of range for database integer (> 2147483647)',
         request: { method: 'POST', url: '/api/sessions/9999999999/certification-candidates' },
       },
+      {
+        condition: 'session ID params is not a number',
+        request: { method: 'GET', url: '/api/sessions/salut/certification-candidates' },
+      },
+      {
+        condition: 'session ID params is out of range for database integer (> 2147483647)',
+        request: { method: 'GET', url: '/api/sessions/9999999999/certification-candidates' },
+      },
     ].forEach(({ condition, request }) => {
       it(`should return 400 when ${condition}`, async function () {
         // given
@@ -133,6 +141,22 @@ describe('Unit | Application | Sessions | Routes', function () {
         // then
         expect(response.statusCode).to.equal(400);
       });
+    });
+  });
+
+  describe('GET /api/sessions/{id}/certification-candidates', function () {
+    it('should exist', async function () {
+      // given
+      sinon.stub(authorization, 'verifySessionAuthorization').returns(null);
+      sinon.stub(certificationCandidateController, 'getCertificationCandidates').returns('ok');
+      const httpTestServer = new HttpTestServer();
+      await httpTestServer.register(moduleUnderTest);
+
+      // when
+      const response = await httpTestServer.request('GET', '/api/sessions/3/certification-candidates');
+
+      // then
+      expect(response.statusCode).to.equal(200);
     });
   });
 });
