+++ api/tests/unit/application/session/index_test.js
@@ -106,22 +106,6 @@ describe('Unit | Application | Sessions | Routes', function () {
     });
   });
 
-  describe('GET /api/sessions/{id}/certification-candidates', function () {
-    it('should exist', async function () {
-      // given
-      sinon.stub(authorization, 'verifySessionAuthorization').returns(null);
-      sinon.stub(sessionController, 'getCertificationCandidates').returns('ok');
-      const httpTestServer = new HttpTestServer();
-      await httpTestServer.register(moduleUnderTest);
-
-      // when
-      const response = await httpTestServer.request('GET', '/api/sessions/3/certification-candidates');
-
-      // then
-      expect(response.statusCode).to.equal(200);
-    });
-  });
-
   describe('DELETE /api/sessions/{id}/certification-candidates/{certificationCandidateId}', function () {
     it('should return 404 if the user is not authorized on the session', async function () {
       // given
@@ -191,14 +175,6 @@ describe('Unit | Application | Sessions | Routes', function () {
       {
         condition: 'session ID params is out of range for database integer (> 2147483647)',
         request: { method: 'GET', url: '/api/admin/sessions/9999999999' },
-      },
-      {
-        condition: 'session ID params is not a number',
-        request: { method: 'GET', url: '/api/sessions/salut/certification-candidates' },
-      },
-      {
-        condition: 'session ID params is out of range for database integer (> 2147483647)',
-        request: { method: 'GET', url: '/api/sessions/9999999999/certification-candidates' },
       },
       {
         condition: 'session ID params is not a number',
